+++ src/org/cicirello/search/operators/permutations/UndoableScrambleMutation.java
@@ -44,12 +44,11 @@ import org.cicirello.math.rand.RandomIndexer;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 4.13.2021
+ * @version 5.13.2021
  */
 public class UndoableScrambleMutation extends Permutation.Mechanic implements UndoableMutationOperator<Permutation> {
 
 	private int[] last;
-	private Permutation previous;
 	private final int[] indexes;
 	
 	/**
@@ -62,7 +61,6 @@ public class UndoableScrambleMutation extends Permutation.Mechanic implements Un
 	@Override
 	public final void mutate(Permutation c) {
 		if (c.length() >= 2) {
-			previous = c;
 			last = c.toArray();
 			generateIndexes(c.length(), indexes);
 			c.scramble(indexes[0], indexes[1]);
@@ -71,9 +69,7 @@ public class UndoableScrambleMutation extends Permutation.Mechanic implements Un
 	
 	@Override
 	public final void undo(Permutation c) {
-		// Verify that c was the most recently mutated permutation.
-		// If so, undo the mutation.
-		if (previous == c) {
+		if (c.length() >= 2) {
 			if (indexes[0] < indexes[1]) {
 				set(c, last, indexes[0], indexes[0], indexes[1]-indexes[0]+1);
 			} else {