+++ src/org/cicirello/search/operators/reals/GaussianMutation.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2020  Vincent A. Cicirello
+ * Copyright (C) 2002-2021  Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  * 
@@ -55,11 +55,9 @@ import org.cicirello.util.Copyable;
  *
  * @param <T> The specific RealValued type.
  *
- * @since 1.0
- *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 6.10.2020
+ * @version 5.12.2021
  */
 public class GaussianMutation<T extends RealValued> implements MutationOperator<T>, RealValued, Copyable<GaussianMutation<T>> {
 	
@@ -165,7 +163,9 @@ public class GaussianMutation<T extends RealValued> implements MutationOperator<
 	 */
 	@Override
 	public boolean equals(Object other) {
-		if (other == null || other.getClass() != getClass()) return false;
+		if (other == null || !(other instanceof GaussianMutation)) {
+			return false;
+		}
 		GaussianMutation g = (GaussianMutation)other;
 		return sigma==g.sigma;
 	}
@@ -288,7 +288,9 @@ public class GaussianMutation<T extends RealValued> implements MutationOperator<
 		 */
 		@Override
 		public boolean equals(Object other) {
-			if (!super.equals(other)) return false;
+			if (!super.equals(other) || !(other instanceof PartialGaussianMutation)) {
+				return false;
+			}
 			PartialGaussianMutation g = (PartialGaussianMutation)other;
 			return k==g.k && p==g.p;
 		}