+++ src/org/cicirello/search/operators/reals/RealVectorInitializer.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2020  Vincent A. Cicirello
+ * Copyright (C) 2002-2021  Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  * 
@@ -37,9 +37,9 @@ import java.util.Arrays;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 9.10.2020
+ * @version 5.12.2021
  */
-public class RealVectorInitializer implements Initializer<RealVector> {
+public final class RealVectorInitializer implements Initializer<RealVector> {
 	
 	private final double[] x;
 	private final double[] a;
@@ -250,10 +250,10 @@ public class RealVectorInitializer implements Initializer<RealVector> {
 	
 	@Override
 	public boolean equals(Object other) {
-		if (other == null || !getClass().equals(other.getClass())) return false;
+		if (other == null || !(other instanceof RealVectorInitializer)) return false;
 		RealVectorInitializer i = (RealVectorInitializer)other;
-		return (min == null && i.min == null ||
-			  Arrays.equals(min, i.min) && Arrays.equals(max, i.max))
+		return ((min == null && i.min == null) ||
+				(Arrays.equals(min, i.min) && Arrays.equals(max, i.max)))
 			  && Arrays.equals(a, i.a) && Arrays.equals(b, i.b);
 	}