+++ src/org/cicirello/search/representations/IntegerVector.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2020  Vincent A. Cicirello
+ * Copyright (C) 2002-2021  Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  * 
@@ -31,7 +31,7 @@ import org.cicirello.util.Copyable;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 6.10.2020
+ * @version 5.12.2021
  */
 public class IntegerVector implements IntegerValued, Copyable<IntegerVector> {
 	
@@ -104,7 +104,9 @@ public class IntegerVector implements IntegerValued, Copyable<IntegerVector> {
 	 */
 	@Override
 	public boolean equals(Object other) {
-		if (other == null || !getClass().equals(other.getClass())) return false;
+		if (other == null || !(other instanceof IntegerVector)) {
+			return false;
+		}
 		return Arrays.equals(x, ((IntegerVector)other).x);
 	}