+++ src/org/cicirello/search/representations/SingleInteger.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2020  Vincent A. Cicirello
+ * Copyright (C) 2002-2021  Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  * 
@@ -26,11 +26,9 @@ import org.cicirello.util.Copyable;
  * A simple class for representing the input to a univariate function, such that
  * the input is an integer.
  *
- * @since 1.0
- *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 6.10.2020
+ * @version 5.12.2021
  */
 public class SingleInteger implements IntegerValued, Copyable<SingleInteger> {
 	
@@ -139,7 +137,7 @@ public class SingleInteger implements IntegerValued, Copyable<SingleInteger> {
 	 */
 	@Override
 	public boolean equals(Object other) {
-		if (other == null || !getClass().equals(other.getClass())) return false;
+		if (other == null || !(other instanceof SingleInteger)) return false;
 		return x == ((SingleInteger)other).x;
 	}