+++ CHANGELOG.md
@@ -4,18 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2021-5-11
+## [Unreleased] - 2021-5-13
 
 ### Added
 
 ### Changed
-* Code improvements (non-functional changes).
+* Code improvements based on report from initial run of MuseDev's 
+  static code analysis tool (non-functional changes to improve code).
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+* Fixed null pointer bug in stochastic sampling implementations for the case when
+  the given heuristic doesn't use an incremental evaluation.
 
 ### CI/CD